convert-cli: rename pipe to pipes and move walk funcs to file
